requirements.txt
autopep8==1.6.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer==2.1.1
ci-info==0.2.0
click==8.1.3
configobj==5.0.6
configparser==5.2.0
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
decorator==5.1.1
dominate==2.5.1
etelemetry==0.3.0
filelock==3.7.1
fitz==0.0.1.dev2
flit-core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
future==0.18.2
httplib2==0.20.4
idna @ file:///croot/idna_1666125576474/work
imageio==2.8.0
importlib-metadata==5.0.0
isodate==0.6.1
Jinja2==3.1.1
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.4.4
looseversion==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.5
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pathlib==1.0.1
Pillow==9.0.1
Polygon3==3.0.8
prov==2.0.0
psutil==5.9.5
pycodestyle==2.9.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pyg-lib @ file:///home/cjc/pytorch1.13.1%2Bcuda116/pyg_lib-0.2.0%2Bpt113cu116-cp37-cp37m-linux_x86_64.whl
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2022.5
pyxnat==1.5
pyzmq==22.3.0
ranger-fm==1.9.3
rdflib==6.2.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn==1.0.2
scipy==1.7.3
Shapely @ file:///tmp/build/80754af9/shapely_1632813481411/work
simplejson==3.17.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.1
torch-cluster @ file:///home/cjc/pytorch1.13.1%2Bcuda116/torch_cluster-1.6.1%2Bpt113cu116-cp37-cp37m-linux_x86_64.whl
torch-geometric==2.2.0
torch-scatter @ file:///home/cjc/pytorch1.13.1%2Bcuda116/torch_scatter-2.1.1%2Bpt113cu116-cp37-cp37m-linux_x86_64.whl
torch-sparse @ file:///home/cjc/pytorch1.13.1%2Bcuda116/torch_sparse-0.6.17%2Bpt113cu116-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ file:///home/cjc/pytorch1.13.1%2Bcuda116/torch_spline_conv-1.2.2%2Bpt113cu116-cp37-cp37m-linux_x86_64.whl
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traits==6.3.2
typing-extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1673575502006/work
visdom==0.2.3
websocket-client==1.3.1
zipp==3.10.0
